Retrieve things that have been flattred from the flattr API. This is done in a separate new minimum priority worker thread FeedStatusFetcher.
We don't provide any user feedback, as this is only a very minor ui-related issue - the list of flattred things is used to disable the flattr menu item as we can't flattr things twice per month.

Mark things as unflattrd if they are not among the things listed by the flattr API. Things where flattring failed (but antennapod thought it worked) are thus re-enabled upon a FlattrStatusFetcher run.

On the other hand, as the URI normalization is tricky and may not cover all the ugly ways in which podcasters or flattr specify URIs, this may yield false negatives
(things marked as unflattrd in AntennaPod that actually have been flattrd).