Parse podcast duration correctly even with trailing whitespace
The Skeptoid podcast (available on gpodder.net) currently has a specific
episode (titled "Skeptoid #437: Tube Amplifiers") that contains the XML
<itunes:duration>14:57 </itunes:duration> which AntennaPod fails
to parse (and instead hits a java.lang.NumberFormatException which
it prints to stderr every time the refresh button is pressed).
The duration for that particular episode is not correctly showed in
the UI, although the duration for other Skeptoid episodes show up
just fine.