Make itunes search more prominent and fix squashed griview
The GridView on large devices (Nexus 6 for example) was showing
iTunes and gpodder results in two columns, which was cutting off
valuable information. I simply double the default size of the grid
columns on the theory that on a tablet someone may still want multiple
columns and that 400dp was wide enough to show the needed detail.

fixes AntennaPod/AntennaPod#1012