Add option to pick lock screen playback buttons
This adds the option to pick which playback buttons to prioritise on the
notification. This allows choosing the playback buttons on the lock
screen. The default playback buttons have not changed and are still set
to play/pause (this is always displayed) and skip.

Note: This commit raises the minimum sdk version from 10 to 11!