Add more playback control options for hardware buttons
The current default behavior of forward and backward hardware buttons
is to fast-foward/rewind, and the user can change this behavior in
preferences by enabling 'Forward button skips' or 'Previous button
restarts'.

This change adds the ability to assign new actions to these hardware
controls. The full set of options is now:

- Skip to the next/previous episode
- Fast forward/Rewind
- Increase/Decrease playback speed
- Restart episode (backward button only)

The default behavior has not changed, the hardware buttons will still
fast-forward/rewind. The older toggle preferences have been removed so
that multiple options could be supported, but if either of them were
enabled the equivalent behavior will be chosen for the new preferences.

In addition, the user can choose to perform any of the custom behaviors
listed above if a button is double-tapped.

Closes #2247