add a "rewind" and "fast forward" button for android Wear
This feature is only enabled for "play" build.
It is useful to skip advertisement during a podcast.

Note: I originaly tried to implement this in the "4 ways controller" managed by
the mediaSession (it contains vol up / vol down / skip). This object is linked
to the capabilities of the session and seems to be hard to customise without
changing the capabilities.
This patch simply add 2 new buttons