Added best-practice checkstyle checks
This are the style guidelines I want to enforce throughout the project in the future.
Currently, huge parts of the app do not pass these checks but when developing something new,
we should add those checks to our IDEs.

They are basically the Google checks but with an increased line length from 100 to 120
characters and using 4 instead of 2 spaces for indentation.