Add playback control options for hardware buttons and repeated actions
The current default behavior of forward and backward hardware buttons
is to fast-forward/rewind, and the user can change this behavior in
preferences by enabling 'Forward button skips' or 'Previous button
restarts'.

This change adds the ability to assign new actions to these hardware
controls. The full set of options is now:

- Skip to the next/previous episode.
- Fast forward/Rewind.
- Increase/Decrease playback speed.
- Restart episode (backward button only).

The default behavior has not changed, the hardware buttons will still
fast-forward/rewind. The older toggle preferences have been removed so
that multiple options could be supported, but if either of them were
enabled the equivalent behavior will be chosen for the new preferences.

In addition, the user can choose to perform any of the custom behaviors
listed above if a button is double-tapped. It also allows the custom
behaviors to be set if the headset button is double or triple-tapped.

The custom repeat action behaviors are only available if the user
enables advanced hardware controls in preferences. Many hardware devices
will have their own custom behaviors for these types of repeated actions
that will prevent the application from even knowing a repeated action
had occured, so the additional preference intends to serve as a
warning the uesr's preference might be ignored.

Implementation notes:

- In order to determine when an external event came from a hardware
  device vs. a notification/widget, a new 'source' extra was added to
  the MediaButtonReceiver intents.
- In order to support skipping to the previous episode in the queue,
  new code paths were added to the playback controller/media player/
  DB tasks.
- Now that changing playback speed can be managed by external key
  events, the logic was moved from AudioplayerActivity to
  PlaybackService.

Closes #2247