Handle HEADSET_HOOK so that it behaves as if MEDIA_PLAY_PAUSE was sent
Without this change, if a HEADSET_HOOK event was sent to the
PlaybackService then it would not be associated with any known
MediaAction. This prevented it from behaving as originally intended,
which is to toggle playback by default or other custom behaviors if
it was sent multiple times and the user had enabled multi-tap events.