Add custom actions for Android Auto
Remove ACTION_SKIP_TO_NEXT from showing because it's hard-coded to
appear on the near left for Android Auto. Replace it with a custom
change playback speed action (not wired up yet) and skip to next
actions in order to fill out the actions list. Add ACTION_PLAY as a
potential action to make sure it shows up in PiP windows.