Add "Skip Inbox" preference
With the "mark as played" swipe action becoming a "played status" toggle
in 2.7.0, I now have to swipe twice on an episode that is in the Inbox
to mark it as played.  Since I do not use the Inbox feature at all, this
has become quite annoying.

This changeset adds a `Skip Inbox` preference which, when turned on,
will prevent "new" episodes from being added to the Inbox.

I also included UI translations for a few languages.