Do not jump bottom sheet when playing
Whenever some view calls requestLayout(), the bottom view jumps.
This happens during slide when setting the player from GONE to VISIBLE.
Also, it happens every time the position changes because the TextView
has a dynamic width. We are not actually interested in the dynamic
width and can simply keep the initial width. This avoids requestLayout()
calls every time the position is updated.