Fix file provider not being able to share all files
The library only looks at one of the external storage devices.
If the exported log file doesn't happen to be stored on
the first device, sharing it doesn't work.
This is a known issue in the Android libraries:
https://issuetracker.google.com/issues/37125252

This commit works around it by using an undocumented element that covers
the entire file system.