Fix file deletion and queueing after download (#6652)
WorkManager does not tell us whether it was cancelled by
the user (not retried) or by the system (retried later).
So we need to delete the file and remove from queue when
we know that it was actually the user. Also make sure
to always delete the file when the download fails.

Also, don't show "will retry" message on last retry attempt.