Move Google Play Metadata to its own repo (#6970)
This avoids having hundreds of megabytes of screenshot data in the main repo every single time we re-generate them. Then developers do not have to clone a huge repo (at least if they clone without submodules). It also enables rewriting the screenshot git history to be smaller without rewriting the code git history (which would be quite a bad idea).