Remove AutoDownloadTest (#7015)
This test regularly fails our CI.

The test checked that auto-download kicks in after the currently playing episode
and that it considers the correct item in the queue to enqueue after.

However, because we now use WorkManager, the download can be delayed based
on decisions by the Android system. We cannot assume that downloading already
starts just seconds after playback completes.

I do not know an easy fix for this, and the test is quite complex anyway, testing
multiple different modules at once. So I am removing the test for now.