Keep track of part sizes while searching atom
This commit introduces a mechanism to keep track of the remaining size of each part while searching for an atom in the M4A file. This is done to ensure that we do not exceed the size of the current part while searching for the next part in the atom name. If the remaining size is exhausted before finding the next part, the search for the atom is aborted and a message is logged. This change improves the robustness of the atom search process and helps prevent potential errors due to exceeding part sizes (e.g. reading the whole file even though the atom is not present).