Fix inconsistent icons in the app toolbar.
The problem was that the icons have their own color they use
(`attr/action_icon_color`) which is fine, except in the toolbar where
they should use a different color.

The fix applies a style to all toolbars where in the style only adds
a theme for the surface which remaps the `attr/action_icon_color` to
the color the other icons in the toolbar use.