Remove duplicated marked as played implementation
There were two implementations, one for single items and one for
multiple items. Now we always use the code for multiple items.

I also made API a lot more simpler and the PodDBAdapter now decides if
a the playback position should be reset. Which it will now do if the
we set it to PLAYED but not for NEW or UNPLAYED. Also it will not reset
an episode that is currently playing.