Fix error icon in the drawer (#7238)
There has been some regression that the error icon in the drawer was
just a circle instead of the the exclamation mark.

The error occurred because the icon was modified to have a background
circle once it was added to the subscription screens. However, in the
drawer a icon tint was applied which now also tinted the background
and made now only the background show.