Refactor ToolbarIconTintManager to a simpler API
The original API used a callback where the caller needed to specify the
icons of the menu items. However, the manager can find them on it's own.

Since the callback was never used for anything else, I also removed it,
which now means the ToolbarIconTintManager can now be used on it's own
instead of instantiating an anonymous class that only overwrites the
callback.