Fixes #2146 Fixes #4535 Fixes #7239 Fixes #7245 Fixes #7246
Reworked sleep timers to allow for clock sleep timers (the original ones) as well as episode sleep timers.
The clock sleep timer works as before, you can configure it to stop playback after a certain amount of time. This time is based on actual seconds, so it is not affected by playback speed.
The clock sleep timer also takes into account pauses, so when the media is paused the sleep timer will no longer advance, it will resume counting down when you start playback again.

The episodes sleep timer allows you to set a timer that stops playback after a certain number of episodes have finished playing. If the option to vibrate is enabled then the episodes sleep timer will vibrate when the last episode starts playing.

If a sleep timer is active then the media duration will always show the remaining time until the sleep timer is hit.
When a sleep timer is active the duration will also have a different color to make it more clear that a sleep timer is active.

When the sleep timer is reset a small vibration will be created so the user has feedback that they've reset the sleep timer duration/episode count.