Sync after refreshing
- Enables users to manually trigger sync
- Makes sure that we actually refresh when a new feed arrives:
Previously, we might request the feed to be refreshed but then don't
actually wait for it to be completed because the refresh service
wouldn't start up quickly enough.

This makes sure that we do not try to sync again before the refresh
actually went through, even if the sync service is called multiple times.