Keep playback service in foreground when media is paused
As reported, pausing media while the notification is visible removes the
playback thread from the foreground state, allowing it to be destroyed
when Android needs more memory.  Some brands of phones seem to reap the
service about a minute after that.

It seems the "play" code behind the notification is not designed to deal
with this removal of the playback service, as playback no longer resumes
when the "play" button is pressed in the notification after the service
has been destroyed.  Unfortunately no clear error appears from the
Android system and all of the logic behind this notification appears to
be abstracted and hidden by `androidx`'s `PlaybackStateCompat`.  After
all, "compact" action buttons for the notification are created with
`PendingIntent.getForegroundService()` which are supposed to (re)start
the service if it went missing.

One solution, proposed here, is to **not** remove the service from the
foreground state when media is simply paused; this allows the playback
button to operate at all times, and is how foreground services are
designed: when the notification is visible, you can be guaranteed that
whatever it is showing/providing is operational.  Technically this
keeps the service in memory longer, but should not cause any increase
in CPU utilization as - if the app is well written... - the service
should be dormant and waiting for its next command (i.e. play) and *not*
spin-looping.

Note that quite a few calls to `stopForeground(false)` remain that could
possibly remove the service from the foreground list *while keeping the
notification visible*.  This seems unintended, but I did not test those
patterns.