Partially work around duplicated feeds when using sync
Connect to each newly added URL with a HEAD request.
If there is a redirect to an existing URL, ignore the added feed.

This does not handle `itunes:new-feed-url` or other redirects.
This will likely still cause problems with synchronization
because the server still knows the old url.
However, it makes the situation better than it was before by avoiding the duplicates.
For a proper fix, we need a new protocol that can deal with redirects more elegantly