Fix removing items from favorites
We check the feed id and the item id when removing from favorites.
However, the item ids are unique, so there is no need to check the feed id.
Actually, I have no idea why it was even added there in the first place.

Now that we have users who somehow ended up with a database where
they cannot un-favorite some of the items, removed the check.

Also make sure to clear the favorites table when deleting a feed.