Work around feeds with broken mime type
If a feed has a broken mime type that is not audio or video,
try to still use it as audio (if there is no audio yet).
This might lead to AP trying to play random other files that are not media files,
but it works around publishers messing up their mime types.