Refactor CrashReportWriter and decouple from the uncaught exception handler.
CrashReportWriter now only handles recording raw exceptions and reading the content of the crash log. It has also been moved to a new "core" module so that it can be consumed by other modules without requiring a dependency on the app module.

A new dedicated UncaughtExceptionHandler class has been added to App module to separate concerns and satisfy the lint 'DefaultUncaughtExceptionDelegation' rule.