Clarify what deleting does (#8165)
We got reports from more than one person that they did not find how to remove a subscription: https://forum.antennapod.org/t/remove-podcast-not-showing-up-on-menu/7779
I think the problem here was that the explanation text does not mention the subscription itself. It only mentions that we remove all episodes.
Update the text to explicitly state that it removes the subscription including its episodes.